# services/facade-api/openapi.yaml
openapi: 3.1.0
info:
  title: Seftec Unified Finance API
  description: |
    Unified API for wallet management, transfers, payments, cards, and KYC verification.

    ## Authentication
    All requests require an API key passed in the `X-API-Key` header.

    ## Idempotency
    For POST requests, include an `Idempotency-Key` header to prevent duplicate operations.
    The key must be unique per request and will be valid for 24 hours.

    ## Webhooks
    Register webhook endpoints to receive real-time notifications for:
    - `transfer.completed`, `transfer.failed`
    - `payment.successful`, `payment.failed`
    - `wallet.credited`, `wallet.debited`
    - `kyc.verified`, `kyc.failed`

    ## Rate Limits
    Rate limits depend on your tier:
    - Starter: 60 req/min, 500 req/hour
    - Growth: 120 req/min, 2000 req/hour
    - Business: 300 req/min, 10000 req/hour
    - Enterprise: Custom

  version: "1.0.0"
  contact:
    name: Seftec API Support
    email: api-support@seftechub.com
    url: https://developers.seftechub.com
  license:
    name: Proprietary
    url: https://seftechub.com/terms
servers:
  - url: https://api.seftechub.com/v1
    description: Production
  - url: https://sandbox.seftechub.com/v1
    description: Sandbox (test mode)
security:
  - ApiKeyAuth: []
...
paths:
...
  /transfers:
    post:
      operationId: createTransfer
      tags: [Transfers]
      summary: Initiate a transfer
      description: |
        Initiate a transfer from a wallet to another wallet or bank account.
        Transfers are processed asynchronously. Use webhooks to receive status updates.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_wallet_id, amount, currency, destination]
              properties:
                source_wallet_id:
                  type: string
                  description: Source wallet ID
                amount:
                  type: number
                  minimum: 1
                  description: Amount in minor units (kobo for NGN)
                currency:
                  $ref: "#/components/schemas/Currency"
                destination:
                  oneOf:
                    - $ref: "#/components/schemas/WalletDestination"
                    - $ref: "#/components/schemas/BankDestination"
                narration:
                  type: string
                  maxLength: 100
                reference:
                  type: string
                  description: Your unique reference (auto-generated if not provided)
                metadata:
                  type: object
            examples:
              wallet_transfer:
                summary: Wallet to Wallet
                value:
                  source_wallet_id: "wal_abc123"
                  amount: 50000
                  currency: "NGN"
                  destination:
                    type: "wallet"
                    wallet_id: "wal_xyz789"
                  narration: "Payment for services"
              bank_transfer:
                summary: Wallet to Bank
                value:
                  source_wallet_id: "wal_abc123"
                  amount: 100000
                  currency: "NGN"
                  destination:
                    type: "bank"
                    account_number: "0123456789"
                    bank_code: "058"
                    account_name: "John Doe"
                  narration: "Withdrawal"
      responses:
        "202":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      operationId: listTransfers
      tags: [Transfers]
      summary: List transfers
      parameters:
        - name: wallet_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TransactionStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferListResponse"
  /transfers/{transfer_id}:
    get:
      operationId: getTransfer
      tags: [Transfers]
      summary: Get transfer status
      parameters:
        - name: transfer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
  /payments:
    post:
      operationId: initializePayment
      tags: [Payments]
      summary: Initialize a payment
      description: |
        Initialize a payment to collect money from a customer.
        Returns payment details including authorization URL for redirect-based payments.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, currency, email]
              properties:
                amount:
                  type: number
                  minimum: 100
                currency:
                  $ref: "#/components/schemas/Currency"
                email:
                  type: string
                  format: email
                reference:
                  type: string
                callback_url:
                  type: string
                  format: uri
                channels:
                  type: array
                  items:
                    type: string
                    enum: [card, bank, ussd, qr, mobile_money, bank_transfer]
                metadata:
                  type: object
            example:
              amount: 250000
              currency: "NGN"
              email: "customer@example.com"
              callback_url: "https://yourapp.com/payment/callback"
              channels: ["card", "bank_transfer"]
      responses:
        "201":
          description: Payment initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /payments/{payment_id}/verify:
    post:
      operationId: verifyPayment
      tags: [Payments]
      summary: Verify a payment
      parameters:
        - name: payment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Your project API key
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Unique key to prevent duplicate operations (recommended for POST requests)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
  schemas:
    Currency:
      type: string
      enum: [NGN, USD, GBP, EUR, GHS, KES, ZAR]
    TransactionStatus:
      type: string
      enum: [pending, processing, success, failed, reversed]
    WalletDestination:
      type: object
      required: [type, wallet_id]
      properties:
        type:
          type: string
          enum: [wallet]
        wallet_id:
          type: string
    BankDestination:
      type: object
      required: [type, account_number, bank_code]
      properties:
        type:
          type: string
          enum: [bank]
        account_number:
          type: string
        bank_code:
          type: string
        account_name:
          type: string
...
    Transfer:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        amount:
          type: number
        fee:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
        status:
          $ref: "#/components/schemas/TransactionStatus"
        source_wallet_id:
          type: string
        destination:
          type: object
        narration:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
        meta:
          type: object
          properties:
            request_id:
              type: string

    # Response wrappers
...
    TransferResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Transfer"
    TransferListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
    PaymentInitResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            reference:
              type: string
            authorization_url:
              type: string
              format: uri
            access_code:
              type: string
            expires_at:
              type: string
              format: date-time
    PaymentVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            reference:
              type: string
            status:
              $ref: "#/components/schemas/TransactionStatus"
            amount:
              type: number
            currency:
              $ref: "#/components/schemas/Currency"
            channel:
              type: string
            paid_at:
              type: string
              format: date-time
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "INVALID_REQUEST"
              message: "Invalid request body"
              details:
                field: "amount"
                issue: "must be positive"
...